Split configuration.py into several files
